rfc3961: Use enctypes, not keytypes
We previously defined the enctype symbols to be aliases for keytype
symbols.  The numerical values matched what we wanted (since these
values are specified in an IANA registry), but the C type is not
required to be the same for enctypes and keytypes.

Some of our buildslave configurations notice the type mismatch and
complain, so fix the types by using the enctype enum for enctype symbols
instead of keytypes.

Change-Id: I56ca634d52954ee44baa34e2d8c876271f171288
Reviewed-on: http://gerrit.openafs.org/10693
Reviewed-by: Derrick Brashear <shadow@your-file-system.com>
Tested-by: Benjamin Kaduk <kaduk@mit.edu>