Revert "viced: Enable NAT ping on hosts"
This reverts commit 53005f4fe7b6df5263d399f0aec9fee3078e39bc.

The fileserver-side "NAT ping" behavior has yet to be proven to be helpful in
situations with NATs. If the behavior is not helpful, this generates
potentially a significant amount of extra useless traffic.  So until it can be
shown to what degree this is helpful, keep this behavior out of the fileserver.

Change-Id: I783222e1be166920445175fe09680cd4968c2810
Reviewed-on: http://gerrit.openafs.org/10135
Reviewed-by: Jeffrey Altman <jaltman@your-file-system.com>
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Stephan Wiesand <stephan.wiesand@desy.de>