libadmin: rework compilation rules
Change I47b3b184a3284b56c4047e27b58b6cb3a8e16cef will add
dependencies on the header files to the library targets. Before
this can be done, we have to avoid the use of $? in compilation
rules because on AIX and OS X this causes the headers to be added
to the compilation command, making builds fail.

This is a 1.6-only change. The corresponding commit on the master
branch, f22ae5f7d8b3c143de21355f819ab872a0cf76fd, is no candidate
for a backport to 1.6.

Change-Id: Iefb432555c60bcfdf51436ebe76a267969280fe0
Reviewed-on: http://gerrit.openafs.org/11353
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Chas Williams - CONTRACTOR <chas@cmf.nrl.navy.mil>
Reviewed-by: D Brashear <shadow@your-file-system.com>
Reviewed-by: Benjamin Kaduk <kaduk@mit.edu>
Reviewed-by: Stephan Wiesand <stephan.wiesand@desy.de>